Move everything around, link underline on hover
