+++ src/app/[elementType]/[elementID]/[[...tab]]/_GeneTabs/_Gene/GeneExpressionUMAP.tsx
@@ -1,36 +1,27 @@
-import { Box, Button, FormControl, InputLabel, MenuItem, Select, SelectChangeEvent, Stack, Typography } from "@mui/material"
+import { GeneExpressionProps, PointMetadata, SharedGeneExpressionPlotProps } from "./GeneExpression"
+import { Box, Button, FormControl, InputLabel, MenuItem, Select, SelectChangeEvent, Typography } from "@mui/material"
 import { getCellCategoryColor, getCellCategoryDisplayname } from "common/utility"
-import { useEffect, useMemo, useRef, useState } from "react"
+import { useMemo, useRef, useState } from "react"
 import { interpolateYlOrRd } from "d3-scale-chromatic";
+import { scaleLinear } from "d3-scale"
 import { Point, ScatterPlot, ChartProps } from "@weng-lab/psychscreen-ui-components"
-import { ParentSize } from "@visx/responsive"
-import { IcreActivityProps, SharedIcreActivityPlotProps, PointMetadata } from "./IcreActivity"
-import { scaleLinear } from "@visx/scale";
 
-export type IcreActivityUmapProps<T, S extends boolean | undefined, Z extends boolean | undefined> =
-  IcreActivityProps
-  & SharedIcreActivityPlotProps
-  & Partial<ChartProps<T, S, Z>>
+export type GeneExpressionUmapProps<T, S extends boolean | undefined, Z extends boolean | undefined> =
+  GeneExpressionProps &
+  SharedGeneExpressionPlotProps &
+  Partial<ChartProps<T, S, Z>>
 
-const IcreActivityUMAP = <T extends PointMetadata, S extends true, Z extends boolean | undefined>({ accession, selected, sortedFilteredData, iCREActivitydata, ...rest }: IcreActivityUmapProps<T, S, Z>) => {
-  const [colorScheme, setColorScheme] = useState<'Zscore' | 'lineage'>('Zscore');
+const GeneExpressionUMAP = <T extends PointMetadata, S extends true, Z extends boolean | undefined>({ geneData, selected, geneExpressionData, ...rest }: GeneExpressionUmapProps<T, S, Z>) => {
+  const [colorScheme, setColorScheme] = useState<'expression' | 'lineage'>('expression');
   const [showLegend, setShowLegend] = useState<boolean>(true);
-  const [assay, setAssay] = useState<'ATAC' | 'DNase' | 'Combined'>('Combined')
 
-  const { data, loading, error } = iCREActivitydata
+  const { data, loading, error } = geneExpressionData
 
   const handleColorSchemeChange = (
     event: SelectChangeEvent,
   ) => {
-    setColorScheme(event.target.value as 'Zscore' | 'lineage');
+    setColorScheme(event.target.value as 'expression' | 'lineage');
   };
-
-  const handleAssayChange = (
-    event: SelectChangeEvent,
-  ) => {
-    setAssay(event.target.value as 'ATAC' | 'DNase' | 'Combined');
-  };
-
   const graphContainerRef = useRef(null);
 
   const map = {
@@ -40,47 +31,29 @@ const IcreActivityUMAP = <T extends PointMetadata, S extends true, Z extends boo
     },
   };
 
-  //prevent scroll on UMAP
-  useEffect(() => {
-    const handleWheel = (event: WheelEvent) => {
-      if (graphContainerRef.current && graphContainerRef.current.contains(event.target)) {
-        event.preventDefault(); // Prevent page scroll
-        // Your zoom logic here
-      }
-    };
-
-    document.addEventListener("wheel", handleWheel, { passive: false });
-
-    return () => {
-      document.removeEventListener("wheel", handleWheel);
-    };
-  }, []);
+  function logTransform(val: number) {
+    return Math.log10(val + 1)
+  }
 
   //find the max logTPM for the domain fo the gradient
   const maxValue = useMemo(() => {
     if (!data || data.length === 0) return 0;
-    return Math.max(...data.map((x) => x.value));
+    return Math.max(...data.map((x) => logTransform(x.value)));
   }, [data]);
-
-
-  //find the max logTPM for the domain fo the gradient
-  const minValue = useMemo(() => {
-    if (!data || data.length === 0) return 0;
-    return Math.min(...data.map((x) => x.value));
-  }, [data]);
-
 
   //generate the domain for the gradient based on the max number
   const generateDomain = (max: number, steps: number) => {
     return Array.from({ length: steps }, (_, i) => (i / (steps - 1)) * max);
   };
 
+  /**
+   * @todo why is this using d3 scaleLinear and not visx
+   */
   const colorScale = useMemo(() =>
-    scaleLinear({
-      domain: generateDomain(maxValue, 9), // 9 evenly spaced domain stops (9 colors)
-      range: Array.from({ length: 9 }, (_, i) => i / 8), // Normalize range for interpolation
-      clamp: true
-    }),
+    scaleLinear<number, number>()
+      .domain(generateDomain(maxValue, 9)) // 9 evenly spaced domain stops (9 colors)
+      .range(Array.from({ length: 9 }, (_, i) => i / 8)) // Normalize range for interpolation
+      .clamp(true),
     [maxValue]
   );
 
@@ -92,29 +65,21 @@ const IcreActivityUMAP = <T extends PointMetadata, S extends true, Z extends boo
   const scatterData: Point<PointMetadata>[] = useMemo(() => {
     if (!data) return []
 
-    const isHighlighted = (d: PointMetadata) => selected.some(x => x.name === d.name)
+    const isHighlighted = (x: PointMetadata) => selected.some(y => y.name === x.name)
 
     return data.map((x) => {
-      const gradientColor = interpolateYlOrRd(colorScale(x.value))
-
-      const getColor = () => {
-        if ((isHighlighted(x) || selected.length === 0)) {
-          if (colorScheme === 'Zscore') {
-            return gradientColor
-          } else return getCellCategoryColor(x.lineage)
-        } else return '#CCCCCC'
-      }
+      const gradientColor = interpolateYlOrRd(colorScale(logTransform(x.value)));
 
       return {
-        x: assay === "Combined" ? x.umap_1 : assay === "ATAC" ? x.umap_atac_1 : x.umap_dnase_1,
-        y: assay === "Combined" ? x.umap_2 : assay === "ATAC" ? x.umap_atac_2 : x.umap_dnase_2,
+        x: x.umap_1,
+        y: x.umap_2,
         r: isHighlighted(x) ? 6 : 4,
-        color: getColor(),
+        color: (isHighlighted(x) || selected.length === 0) ? ((colorScheme === 'expression') ? gradientColor : getCellCategoryColor(x.lineage)) : '#CCCCCC',
         shape: x.stimulation === "unstimulated" ? "circle" : "triangle" as "circle" | "triangle",
         metaData: x
       };
     }).sort((a, b) => (isHighlighted(b.metaData)) ? -1 : 0)
-  }, [data, colorScale, selected, assay, colorScheme]);
+  }, [data, colorScale, selected, colorScheme]);
 
   const legendEntries = useMemo(() => {
     if (!scatterData) return [];
@@ -140,36 +105,15 @@ const IcreActivityUMAP = <T extends PointMetadata, S extends true, Z extends boo
       <>
         <Typography><b>Lineage:</b> {getCellCategoryDisplayname(point.metaData.lineage)}</Typography>
         <Typography><b>Biosample:</b> {point.metaData.biosample}, {point.metaData.stimulation}</Typography>
-        <Typography><b>Assay:</b> {point.metaData.assay}</Typography>
-        <Typography><b>Z-score:</b> {point.metaData.value.toFixed(2)}</Typography>
+        <Typography><b>TPM:</b> {point.metaData.value.toFixed(1)}</Typography>
         <Typography><b>Source:</b> {point.metaData.source}</Typography>
       </>
     )
   }
 
-  const AssaySelect = () => {
-    return (
-      <FormControl>
-        <InputLabel>Assay</InputLabel>
-        <Select
-          labelId="demo-simple-select-label"
-          id="demo-simple-select"
-          value={assay}
-          label="Assay"
-          onChange={handleAssayChange}
-          MenuProps={{ disableScrollLock: true }}
-        >
-          <MenuItem value={"Combined"}>ATAC & DNase</MenuItem>
-          <MenuItem value={"ATAC"}>ATAC</MenuItem>
-          <MenuItem value={"DNase"}>DNase</MenuItem>
-        </Select>
-      </FormControl>
-    )
-  }
-
   const ColorBySelect = () => {
     return (
-      <FormControl>
+      <FormControl sx={{ alignSelf: "flex-start" }}>
         <InputLabel>Color By</InputLabel>
         <Select
           labelId="demo-simple-select-label"
@@ -179,7 +123,7 @@ const IcreActivityUMAP = <T extends PointMetadata, S extends true, Z extends boo
           onChange={handleColorSchemeChange}
           MenuProps={{ disableScrollLock: true }}
         >
-          <MenuItem value={"Zscore"}>Z-score</MenuItem>
+          <MenuItem value={"expression"}>Expression</MenuItem>
           <MenuItem value={"lineage"}>Lineage</MenuItem>
         </Select>
       </FormControl>
@@ -188,17 +132,15 @@ const IcreActivityUMAP = <T extends PointMetadata, S extends true, Z extends boo
 
   return (
     <>
-      <Stack direction={"row"} spacing={2}>
-        <ColorBySelect />
-        <AssaySelect />
-      </Stack>
-      <Box
-        padding={1}
+      <ColorBySelect />
+      <Box 
+        padding={1} 
         //hacky height, have to subtract the pixel value of the Colorby select and the margin to line it up with the table
-        sx={{ border: '1px solid', borderColor: 'divider', borderRadius: 1, position: "relative", width: "100%", height: "calc(100% - 72px)" }}
-        ref={graphContainerRef}
-        mt={2}
+        sx={{ border: '1px solid', borderColor: 'divider', borderRadius: 1, position: "relative", width: "100%", height: "calc(100% - 72px)" }} 
+        ref={graphContainerRef} 
+        mt={2} 
         mb={2}
+        zIndex={0}
       >
         <Typography variant="body2" align="right">
           {"\u25EF unstimulated, \u25B3 stimulated "}
@@ -211,7 +153,8 @@ const IcreActivityUMAP = <T extends PointMetadata, S extends true, Z extends boo
           loading={loading}
           miniMap={map}
           groupPointsAnchor="lineage"
-          tooltipBody={(point) => <TooltipBody {...point} />
+          tooltipBody={(point) => <TooltipBody {...point}
+          />
           }
         />
         <Button variant="outlined" sx={{ position: "absolute", bottom: 10, left: 10, textTransform: "none" }} onClick={() => setShowLegend(!showLegend)}>Toggle Legend</Button>
@@ -220,11 +163,11 @@ const IcreActivityUMAP = <T extends PointMetadata, S extends true, Z extends boo
       {showLegend && (
         <Box sx={{ display: 'flex', flexDirection: 'column' }}>
           <Typography mb={1}><b>Legend</b></Typography>
-          {colorScheme === "Zscore" ? (
+          {colorScheme === "expression" ? (
             <>
-              <Typography>Z-score</Typography>
+              <Typography>Log₁₀(TPM + 1)</Typography>
               <Box sx={{ display: "flex", alignItems: "center", width: "200px" }}>
-                <Typography sx={{ mr: 1 }}>{'< 0'}</Typography>
+                <Typography sx={{ mr: 1 }}>0</Typography>
                 <Box
                   sx={{
                     height: "16px",
@@ -266,4 +209,4 @@ const IcreActivityUMAP = <T extends PointMetadata, S extends true, Z extends boo
   )
 }
 
-export default IcreActivityUMAP
+export default GeneExpressionUMAP
