+++ src/app/[elementType]/[elementID]/[[...tab]]/_SnpTabs/_Variant/SnpFrequencies.tsx
