+++ src/app/[elementType]/[elementID]/[[...tab]]/_SnpTabs/_Variant/Variant.tsx
@@ -1,0 +1,14 @@
+import { Stack } from "@mui/material";
+import SnpGWASLdr from "./SnpGWASLdr";
+import SnpFrequencies from "./SnpFrequencies";
+
+const VariantInfo = ({ snpid }: { snpid: string }) => {
+  return (
+    <Stack spacing={2}>
+      <SnpFrequencies snpid={snpid} />
+      <SnpGWASLdr snpid={snpid} />{" "}
+    </Stack>
+  );
+};
+
+export default VariantInfo;
