+++ src/common/components/VerticalBarPlot.tsx
