+++ src/common/utility.tsx
@@ -1,12 +1,6 @@
 import { GenomicRange } from "types/globalTypes"
 import { cellCategoryColors, cellCategoryDisplaynames, studyLinks } from "./consts"
-import { Launch } from "@mui/icons-material";
-import { Link as MuiLink, LinkProps as MuiLinkProps, Typography, TypographyOwnProps, TypographyPropsVariantOverrides } from "@mui/material";
-import NextLink from 'next/link'
-import { LinkProps as NextLinkProps } from "next/link";
-/**
- * @todo Merge with utility.ts
- */
+import { Typography, TypographyOwnProps } from "@mui/material";
 
 export function getClassDisplayname(input: string) {
   switch (input) {
@@ -24,9 +18,9 @@ export function getClassDisplayname(input: string) {
 }
 
 /**
- * 
- * @param input chr:start-end
- * @returns Very dumb parser for genomic range. No input checking. Assumes proper formatting and no commas in values
+ * Very dumb parser for genomic range. No input checking. Assumes proper formatting and no commas in values
+ * @param input `String` with format chr:start-end
+ * @returns `GenomicRange`
  */
 export function parseGenomicRangeString(input: string): GenomicRange {
   if (input.includes("%3A")) {
@@ -83,39 +77,6 @@ export function getCellCategoryDisplayname(cell: string) {
 export function getStudyLink(study: string) {
   return studyLinks[study] || "Unknown Study"
 }
-
-//Combines the props of Next Link and Mui Link, adds the custom props
-export type LinkComponentProps = Omit<MuiLinkProps, "component" | "href"> &
-  Omit<NextLinkProps, "passHref"> & {
-    showExternalIcon?: boolean;
-    openInNewTab?: boolean;
-  };
-
-/**
- * Mui Link styled component, Next Link behavior
- * @param props ```MuiLinkProps & { showExternalIcon?: boolean }```
- * @returns 
- */
-export const LinkComponent = ({
-  showExternalIcon,
-  openInNewTab = false,
-  children,
-  ...rest
-}: LinkComponentProps) => {
-  return (
-    <MuiLink
-      component={NextLink}
-      rel={openInNewTab ? "noopener noreferrer" : undefined}
-      target={openInNewTab ? "_blank" : undefined}
-      {...rest}
-    >
-      {children}
-      {showExternalIcon && (
-        <Launch sx={{ display: "inline-flex", verticalAlign: "middle", ml: 0.5 }} color="inherit" fontSize="inherit" />
-      )}
-    </MuiLink>
-  );
-};
 
 /**
  * Only use this function if use case is unable to handle jsx element and needs string. Use `toScientificNotationElement` if possible
